+++ crates/build_safely/examples/unstable/default_field_values/build.rs
@@ -1,0 +1,8 @@
+use build_safely::prelude::*;
+
+fn main() -> Result<()> {
+    let mut ac = AutoCfg::new()?;
+    let allowed_features = cargo_allowed_features()?;
+    ac.emit_unstable_feature(default_field_values, &allowed_features);
+    Ok(())
+}