+++ crates/build_safely/examples/unstable/negative_impls/src/lib.rs
@@ -1,0 +1,20 @@
+#![cfg_attr(unstable_negative_impls, feature(negative_impls))]
+#![allow(unused)]
+
+#[cfg(has_negative_impls)]
+mod has {
+    #[test]
+    fn has() {
+        struct Prisoner;
+        impl !Send for Prisoner {}
+    }
+}
+
+#[cfg(not(has_negative_impls))]
+mod has_not {
+    /// ```compile_fail
+    /// struct Prisoner;
+    /// impl !Send for Prisoner {}
+    /// ```
+    fn doctest() {}
+}