+++ crates/build_safely/examples/unstable/try_v2/build.rs
@@ -1,7 +1,7 @@
 use build_safely::prelude::*;
 
 fn main() -> Result<()> {
-    let ac = AutoCfg::new()?;
+    let mut ac = AutoCfg::new()?;
     let allowed_features = cargo_allowed_features()?;
     ac.emit_unstable_feature(try_trait_v2, &allowed_features);
     ac.emit_unstable_feature(try_trait_v2_residual, &allowed_features);